rename second LobbyService to WebsocketLobbyService and cleanup dependencies
